Login

+++ app/Http/Controllers/TasksController.php
@@ -15,11 +15,18 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
-
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+        $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+            
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
+        
+        return view('welcome', $data);
     }
 
     /**
@@ -48,10 +55,10 @@ class TasksController extends Controller
             'status' => 'required|max:10', 
             'content' => 'required|max:191',
         ]);
-        $task = new Task;
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
+        $request->user()->tasks()->create([
+            'content' =>$request->content,
+            'status' =>$request->status,
+        ]);
 
         return redirect('/');
     }
@@ -64,6 +71,10 @@ class TasksController extends Controller
     public function show($id)
     {
         $task = Task::find($id);
+        
+        if (\Auth::id() !== $task->user_id) {
+            return redirect('/');
+        }
 
         return view('tasks.show', [
             'task' => $task,
@@ -79,6 +90,9 @@ class TasksController extends Controller
     public function edit($id)
     {
         $task = Task::find($id);
+        if (\Auth::id() !== $task->user_id) {
+            return redirect('/');
+        }
 
         return view('tasks.edit', [
             'task' => $task,
@@ -99,6 +113,10 @@ class TasksController extends Controller
             'content' => 'required|max:191',
         ]);
         $task = Task::find($id);
+        if (\Auth::id() !== $task->user_id) {
+            return redirect('/');
+        }
+        
         $task->status = $request->status;
         $task->content = $request->content;
         $task->save();
@@ -115,6 +133,9 @@ class TasksController extends Controller
     public function destroy($id)
     {
         $task = Task::find($id);
+        if (\Auth::id() !== $task->user_id) {
+            return redirect('/');
+        }
         $task->delete();
 
         return redirect('/');
